Struct dynlink::context::Context

#[repr(C)]
pub struct Context<Engine: ContextEngine> { /* private fields */ }

A dynamic linker context, the main state struct for this crate.

Implementations

impl<Engine: ContextEngine> Context<Engine>

pub fn load_library_in_compartment<Namer>( &mut self, comp_id: CompartmentId, unlib: UnloadedLibrary, namer: Namer ) -> Result<LibraryId, DynlinkError>where Namer: FnMut(&str) -> Option<Engine::Backing> + Clone,

Load a library into a given compartment. The namer callback resolves names to Backing objects, allowing the caller to hook into the “name-of-dependency” -> backing object pipeline.

impl<Engine: ContextEngine> Context<Engine>

pub fn relocate_all(&mut self, root_id: LibraryId) -> Result<(), DynlinkError>

Iterate through all libraries and process relocations for any libraries that haven’t yet been relocated.

impl<Engine: ContextEngine> Context<Engine>

pub fn build_ctors_list( &self, root_id: LibraryId ) -> Result<Vec<CtorInfo>, DynlinkError>

Build up a list of constructors to call for a library and its dependencies.

pub fn build_runtime_info( &self, root_id: LibraryId, tls: TlsRegion ) -> Result<RuntimeInitInfo, DynlinkError>

Build the runtime handoff info for bootstrapping the Twizzler runtime.

impl<Engine: ContextEngine> Context<Engine>

pub fn lookup_symbol<'a>( &'a self, start_id: LibraryId, name: &str, lookup_flags: LookupFlags ) -> Result<RelocatedSymbol<'a, Engine::Backing>, DynlinkError>

Search for a symbol, starting from library denoted by start_id. For normal symbol lookup, this should be the ID of the library that needs a symbol looked up. Flags can be specified which allow control over where to look for the symbol.

impl<Engine: ContextEngine> Context<Engine>

pub fn new(engine: Engine) -> Self

Construct a new dynamic linker context.

pub fn lookup_compartment(&self, name: &str) -> Option<CompartmentId>

Lookup a compartment by name.

pub fn get_compartment( &self, id: CompartmentId ) -> Result<&Compartment<Engine::Backing>, DynlinkError>

Get a reference to a compartment back by ID.

pub fn get_compartment_mut( &mut self, id: CompartmentId ) -> Result<&mut Compartment<Engine::Backing>, DynlinkError>

Get a mut reference to a compartment back by ID.

pub fn lookup_library( &self, comp: CompartmentId, name: &str ) -> Option<LibraryId>

Lookup a library by name

pub fn get_library( &self, id: LibraryId ) -> Result<&Library<Engine::Backing>, DynlinkError>

Get a reference to a library back by ID.

pub fn get_library_mut( &mut self, id: LibraryId ) -> Result<&mut Library<Engine::Backing>, DynlinkError>

Get a mut reference to a library back by ID.

pub fn with_dfs_postorder<R>( &self, root_id: LibraryId, f: impl FnMut(&LoadedOrUnloaded<Engine::Backing>) -> R ) -> Vec<R>

Traverse the library graph with DFS postorder, calling the callback for each library.

pub fn with_dfs_postorder_mut<R>( &mut self, root_id: LibraryId, f: impl FnMut(&mut LoadedOrUnloaded<Engine::Backing>) -> R ) -> Vec<R>

Traverse the library graph with DFS postorder, calling the callback for each library (mutable ref).

pub fn with_bfs( &self, root_id: LibraryId, f: impl FnMut(&LoadedOrUnloaded<Engine::Backing>) )

Traverse the library graph with BFS, calling the callback for each library.

pub fn libraries(&self) -> LibraryIter<'_, Engine>

pub fn add_manual_dependency(&mut self, parent: LibraryId, dependee: LibraryId)

pub fn add_compartment( &mut self, name: impl ToString ) -> Result<CompartmentId, DynlinkError>

Create a new compartment with a given name.