dynlink::engines

Trait ContextEngine

pub trait ContextEngine {
    // Required methods
    fn load_segments(
        &mut self,
        src: &Backing,
        ld: &[LoadDirective],
        comp_id: CompartmentId,
        load_ctx: &mut LoadCtx,
    ) -> Result<Vec<Backing>, DynlinkError>;
    fn load_object(
        &mut self,
        unlib: &UnloadedLibrary,
    ) -> Result<Backing, DynlinkError>;
    fn select_compartment(
        &mut self,
        unlib: &UnloadedLibrary,
    ) -> Option<CompartmentId>;

    // Provided methods
    fn add_name_map(&mut self, _name: &str, _id: ObjID) { ... }
    fn remove_name_map(&mut self, _name: Option<&str>, _id: Option<ObjID>) { ... }
}

System-specific implementation functions for the dynamic linker, mostly involving loading objects.

Required Methods

fn load_segments( &mut self, src: &Backing, ld: &[LoadDirective], comp_id: CompartmentId, load_ctx: &mut LoadCtx, ) -> Result<Vec<Backing>, DynlinkError>

Load a given source backing into new backings, according to the given load directives.

fn load_object( &mut self, unlib: &UnloadedLibrary, ) -> Result<Backing, DynlinkError>

Load a single object, based on the given unloaded library.

fn select_compartment( &mut self, unlib: &UnloadedLibrary, ) -> Option<CompartmentId>

Select which compartment a library should go in.

Provided Methods

fn add_name_map(&mut self, _name: &str, _id: ObjID)

fn remove_name_map(&mut self, _name: Option<&str>, _id: Option<ObjID>)