Trait dynlink::library::BackingData

pub trait BackingData: Clone {
    type InnerType;

    // Required methods
    fn data(&self) -> (*mut u8, usize);
    fn load_addr(&self) -> usize;
    fn to_inner(self) -> Self::InnerType;
    fn inner(&self) -> &Self::InnerType;

    // Provided methods
    fn slice(&self) -> &[u8]  { ... }
    fn get_elf(&self) -> Result<ElfBytes<'_, NativeEndian>, ParseError> { ... }
}

The core trait that represents loaded or mapped data.

Required Associated Types

type InnerType

Required Methods

fn data(&self) -> (*mut u8, usize)

Get a pointer to the start of a region, and a length, denoting valid memory representing this object. The memory region is valid.

fn load_addr(&self) -> usize

fn to_inner(self) -> Self::InnerType

Get the inner implementation type.

fn inner(&self) -> &Self::InnerType

Provided Methods

fn slice(&self) -> &[u8]

Get the data as a slice.

fn get_elf(&self) -> Result<ElfBytes<'_, NativeEndian>, ParseError>

Get the ELF file for this backing.