dynlink::context

Struct Context

#[repr(C)]
pub struct Context { /* private fields */ }

A dynamic linker context, the main state struct for this crate.

Implementations

impl Context

pub fn load_library_in_compartment( &mut self, comp_id: CompartmentId, unlib: UnloadedLibrary, allowed_gates: AllowedGates, load_ctx: &mut LoadCtx, ) -> Result<Vec<LoadIds>, DynlinkError>

Load a library into a given compartment.

impl Context

pub fn relocate_all(&mut self, root_id: LibraryId) -> Result<(), DynlinkError>

Iterate through all libraries and process relocations for any libraries that haven’t yet been relocated.

impl Context

pub fn build_ctors_list( &self, root_id: LibraryId, comp: Option<CompartmentId>, ) -> Result<Vec<CtorSet>, DynlinkError>

Build up a list of constructors to call for a library and its dependencies.

pub fn build_runtime_info( &self, root_id: LibraryId, tls: TlsRegion, ) -> Result<RuntimeInitInfo, DynlinkError>

Build the runtime handoff info for bootstrapping the Twizzler runtime.

impl Context

pub fn lookup_symbol<'a>( &'a self, start_id: LibraryId, name: &str, lookup_flags: LookupFlags, ) -> Result<RelocatedSymbol<'a>, DynlinkError>

Search for a symbol, starting from library denoted by start_id. For normal symbol lookup, this should be the ID of the library that needs a symbol looked up. Flags can be specified which allow control over where to look for the symbol.

impl Context

pub fn new(engine: Box<dyn ContextEngine + Send>) -> Self

Construct a new dynamic linker context.

pub fn replace_engine(&mut self, engine: Box<dyn ContextEngine + Send>)

Replace the callback engine for this context.

pub fn lookup_compartment(&self, name: &str) -> Option<CompartmentId>

Lookup a compartment by name.

pub fn get_compartment( &self, id: CompartmentId, ) -> Result<&Compartment, DynlinkError>

Get a reference to a compartment back by ID.

pub fn get_compartment_mut( &mut self, id: CompartmentId, ) -> Result<&mut Compartment, DynlinkError>

Get a mut reference to a compartment back by ID.

pub fn lookup_library( &self, comp: CompartmentId, name: &str, ) -> Option<LibraryId>

Lookup a library by name

pub fn get_library(&self, id: LibraryId) -> Result<&Library, DynlinkError>

Get a reference to a library back by ID.

pub fn get_library_mut( &mut self, id: LibraryId, ) -> Result<&mut Library, DynlinkError>

Get a mut reference to a library back by ID.

pub fn with_dfs_postorder<R>( &self, root_id: LibraryId, f: impl FnMut(&LoadedOrUnloaded) -> R, ) -> Vec<R>

Traverse the library graph with DFS postorder, calling the callback for each library.

pub fn with_dfs_postorder_mut<R>( &mut self, root_id: LibraryId, f: impl FnMut(&mut LoadedOrUnloaded) -> R, ) -> Vec<R>

Traverse the library graph with DFS postorder, calling the callback for each library (mutable ref).

pub fn with_bfs( &self, root_id: LibraryId, f: impl FnMut(&LoadedOrUnloaded) -> bool, )

Traverse the library graph with BFS, calling the callback for each library.

pub fn libraries(&self) -> LibraryIter<'_>

pub fn add_manual_dependency(&mut self, parent: LibraryId, dependee: LibraryId)

pub fn unload_compartment( &mut self, comp_id: CompartmentId, ) -> (Option<Compartment>, Vec<LoadedOrUnloaded>)

pub fn add_compartment( &mut self, name: impl ToString, new_comp_flags: NewCompartmentFlags, ) -> Result<CompartmentId, DynlinkError>

Create a new compartment with a given name.

pub fn compartment_dependencies( &self, id: CompartmentId, ) -> Result<Vec<CompartmentId>, DynlinkError>

Get a list of external compartments that the given compartment depends on.